Added authorisation checks

+++ app/Http/Controllers/PostController.php
@@ -15,7 +15,8 @@ class PostController extends Controller
      */
     public function index()
     {
-      $posts = Post::orderBy('updated_at', 'desc')->get();
+
+      $posts = Post::where('user_id', auth()->id())->orderBy('updated_at', 'desc')->get();
       return view('posts.index', compact('posts'));
     }
 
@@ -43,10 +44,10 @@ class PostController extends Controller
 
       Post::create([
         'body' =>  request('body'),
-        'user_id' => \App\User::first()->id
+        'user_id' => auth()->id()
       ]);
 
-      return redirect('posts');
+      return redirect('/');
     }
 
     /**
@@ -68,6 +69,7 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        abort_unless(auth()->user()->owns($post), 403);
         return view('posts.edit', compact('post'));
     }
 
@@ -80,13 +82,14 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        abort_unless(auth()->user()->owns($post), 403);
         request()->validate([
           'body' => 'required'
         ]);
 
         $post->update(request(['body']));
 
-        return redirect('posts');
+        return redirect('/');
     }
 
     /**
@@ -97,8 +100,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_unless(auth()->user()->owns($post), 403);
         $post->delete();
 
-        return redirect('posts');
+        return redirect('/');
     }
 }
